requirements.txt
Django==1.6.5
South==1.0
dj-database-url==0.3.0
dj-static==0.0.5
django-model-utils==2.0.3
django-notifications-hq==0.6.2
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
wsgiref==0.1.2
django-filter==0.7
djangorestframework==2.3.14
